+++ rolling/base.py
@@ -1,4 +1,5 @@
 import abc
+from itertools import chain
 
 
 class RollingObject(metaclass=abc.ABCMeta):
@@ -45,7 +46,7 @@ class RollingObject(metaclass=abc.ABCMeta):
         self = super().__new__(cls)
 
         self.window_type = window_type
-        self.window_size = self._validate_window_size(window_size)
+        self.window_size = _validate_window_size(window_size)
         self._iterator = iter(iterable)
 
         if self.window_type == "variable":
@@ -95,6 +96,25 @@ class RollingObject(metaclass=abc.ABCMeta):
                 self._remove_old()
                 return self.current_value
 
+    def extend(self, iterable):
+        """
+        Extend the iterator being consumed with a new iterable.
+
+        The extend() method may be called at any time (even after
+        StopIteration has been raised). The most recent values from
+        the current iterator are retained and used in the calculation
+        of the next window value.
+
+        For "variable" windows which are decreasing in size, extending
+        the iterator means that these windows will grow towards their
+        maximum size again.
+
+        """
+        self._iterator = chain(self._iterator, iterable)
+
+        if self.window_type == "variable":
+            self._filled = False
+
     @property
     @abc.abstractmethod
     def current_value(self):
@@ -138,15 +158,15 @@ class RollingObject(metaclass=abc.ABCMeta):
         """
         pass
 
-    @staticmethod
-    def _validate_window_size(k):
-        """
-        Check if k is a positive integer
-        """
-        if not isinstance(k, int):
-            raise TypeError(
-                "window_size must be integer type, got {}".format(type(k).__name__)
-            )
-        if k <= 0:
-            raise ValueError("window_size must be positive")
-        return k
+
+def _validate_window_size(k):
+    """
+    Check if k is a positive integer
+    """
+    if not isinstance(k, int):
+        raise TypeError(
+            "window_size must be integer type, got {}".format(type(k).__name__)
+        )
+    if k <= 0:
+        raise ValueError("window_size must be positive")
+    return k